vanna-ai/requirements.txt
vanna[postgres]
flask
flask-cors
python-dotenv
psycopg2-binary
